workingEnvironment/03_network/NETS_MI_PVAL_v2/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "mine-network"
version = "0.1.0"
description = "MINE-based gene co-expression network inference with permutation significance testing, multi-study consensus, MCODE module detection, and gene-set annotation."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "DSL Course"},
]

dependencies = [
    "numpy>=1.21",
    "pandas>=1.3",
    "scipy>=1.7",
    "torch>=1.10",
    "joblib>=1.0",
    "python-igraph>=0.9",
    "matplotlib>=3.4",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[project.scripts]
mine-network = "mine_network.pipeline:run_pipeline"

[tool.setuptools.packages.find]
include = ["mine_network*"]
